+++ CHANGELOG.md
@@ -3,6 +3,12 @@
 This project adheres to [Semantic Versioning](http://semver.org/).
 
 ## [unreleased]
+- Feature: Add new `k-RadioButton` React component.
+- Feature: Add new `k-SideGrid` component.
+- Fix: Change `k-SingleColumn` width according to the media.
+
+- Feature: Add new `SideGrid`, `SideGridContent` and `SideGridAside` components.
+- Feature: Add new `upcaseFirst()` string helper.
 
 - Feature: Add new `initialTouched` and `requiredDurationError` props
   to `LoanSimulator`.